# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skillup API",
    "version": "1.0.0",
    "description": "API documentation for Skillup Learning Management System"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/assignments": {
      "post": {
        "summary": "Create a new assignment (Tutor only)",
        "tags": [
          "Assignments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all assignments",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/assignments/{assignmentId}/submit": {
      "post": {
        "summary": "Submit work for an assignment (Learner only)",
        "tags": [
          "Submissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/assignments/{assignmentId}/submissions": {
      "get": {
        "summary": "Get all submissions for an assignment (Tutor only)",
        "tags": [
          "Submissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/submissions/{submissionId}/review": {
      "post": {
        "summary": "Review and grade a submission (Tutor only)",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}